requirements.txt
PyPDF2
nltk
torch
transformers
chromadb
ollama
redis
faiss-cpu
preprocess
psutil